this adds few more screens to the routes, to start implement login and home
