+++ lib/MessageUtils.dart
@@ -1,0 +1,5 @@
+class MessageUtils {
+  List<String> salute = [
+    'Oi, meu nome é Débora da bico, hoje vou te ajudar com seu cadastro, como você está?'
+  ];
+}
