+++ lib/utils.dart
@@ -2,6 +2,8 @@ import 'dart:io';
 
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+import 'package:flutter_statusbarcolor/flutter_statusbarcolor.dart';
 import 'package:path/path.dart' as path;
 import 'package:tinycolor/tinycolor.dart';
 
@@ -35,6 +37,12 @@ class Utils {
 
   static RegExp intRegex() {
     return RegExp(r'[^0-9]');
+  }
+
+  static setBarAndNavColor(BuildContext context) {
+    FlutterStatusbarcolor.setStatusBarColor(Utils.barcolor(context));
+    SystemChrome.setSystemUIOverlayStyle(SystemUiOverlayStyle(
+        systemNavigationBarColor: Theme.of(context).scaffoldBackgroundColor));
   }
 
   static Color barcolor(BuildContext context) {
