+++ pubspec.yaml
@@ -32,6 +32,7 @@ dependencies:
   email_validator: ^1.0.0
   cpf_cnpj_validator: ^1.0.5
   tinycolor: ^1.0.2
+  flutter_statusbarcolor: ^0.2.3
 
 
 flutter_icons:
